Complete initial implementation of string-to-set
Testing this is going to have to wait on writing the function
which can test set membership, after which it's easy: make some
strings, feed their mutable copy in, then progressively iterate
the string and confirm membership.

Since we'll have plenty of string around, with non-overlapping
magisteria, we can at least do some sanity checks against non-
membership as well.

Testing that all erroneous input is handled correctly is a job for
a fuzzer.  So that's a thing to figure out.