Test for four-byte charset
With the usual bugs shaken out in the process.

There are likely some untested code paths in here which I won't
see until we add a wider spectrum of characters, but I caught
myself using Julia indexing in this round and I think I fixed all
the spots which had it, so it might even hold up under load already.

Next up is the really mind-bending stuff: union, intersection, and
difference.  I have the algorithms worked out, but text is text and
it's code which speaks.