Start RuneSet type
Decided to support only one sort of RuneSet.  For one thing, the
representation of RuneSets is inherently dense when the set is
dense in Unicode space, so the savings are not very great.

For another, blocks (inherently a dense range) are less interesting
than scripts, which more often then not have a few gaps in the
range which we don't want to spuriously match.

Since we'll also be building RuneSets from ranges, it will make
sense to have a distinct UnicodeRange type, which will also provide
a membership test.  But they'll be separate types, with conversion
to RuneSet provided so that set operations may be performed
without trying to break my skull against doing it directly.