Add T4 rank of set union
This is all rather intricate and it would be astonishing if it
worked right out of the gate.  Worth refactoring this into some
helper functions and thinking about ways to simplify some of the
tracking logic, before tackling difference and intersection.

Oh and definitely want to figure out how to hook up the VSCode
debugger...