Feathered four-byte setunion works correctly
Problem turned out to be the logic of the reverse iterator.

Rewriting it to use an i8 and an `@intCast` solved the problem, I
just hope that the current compiler can generate optimal code for
this case.

If necessary, going back and manually applying backwards iteration
in the key set operation points which use it is at least practical.

Then again, I want set operations to be efficient, but it's likely
that the only hot loop in a program will be membership.  Even if
we have some less than optimal code, it's reasonable to wait around
and see if the compiler catches up. All this is speculative anyhow.