Complete rewrite of T3 level of set intersection
A mentally fatiguing exercise, to put it mildly.

Confident that the fundamental logic is correct this time, well
worth putting a pin in it and generating some tests of said logic
before proceeding to the T4 tier.

It's the same thing with an additional level of nesting, making it
*relatively* straightforward, but that's not to say I'm looking
forward to actually writing it.

Intersection will be easier, because we can eliminate lead bytes
right away.  Though we do have to circle back and stub out more
later, so hmm. Maybe not.