Change receiver to ordinary reference
Realized there's no compelling reason to use a * const for the
RuneSet receivers.  Need to read up at some point on what the
difference really is, but I think the main effect there would be
to constrain the compiler in how it approaches the function call,
and this is (probably!) not the right thing to do.