Remove redundant checks for set having c or d byte
Every set has a LEAD mask, so we don't have to worry about going
past the set slice bounds if we get a codeunit with c or d bytes
when the set doesn't have a T3 or T4.  The membership check for
set[LEAD] will reject all such codepoints, provided that the set
itself is properly constructed.