Scatter data suggests T3 ops are clean
The most likely explanation for all of this, hilariously, is that
I'm iterating the T3-T4 boundary backward, in all three data sets.

This was concealed by using data which was suspiciously regular,
to the point where it *didn't matter which order* the final masks
were in.

There was a reason for that: I wanted data which would exercise
the full range of iterators, and knock out entire blocks at a
time.  But then I wrote tests which were supposed to knock out
everything.  Anyway, there was always going to be more data, and
the order in which I catch bugs is not the most important part of
development.  There's still a lesson in here somewhere.