Set intersection tests pass
I'll be adding more extensive tests, but I expect any remaining
bugs to be minor.

The library isn't quite feature complete, I do want to be able to
serialize the string corresponding to the set, there may be some
other minor operations to mop up.  But the job is largely done.
Huzzah.