Test refactor
This combines all the separate operations tests into a single
interface, so that I can add test data and apply all the validity
tests to it with a single line.

As of this commit, runeset.zig is at 97.9% coverage.  There's an
assert in setDifference, and a branch in setIntersection, which I
haven't figured out how to trigger, and the set creation function
needs to be exercised against invalid Unicode data to hit the
bailout branches.

As nice as 100% coverage will be to reach, it doesn't guarantee
that every line of each algorithm is valid.  This can't, in fact,
be guaranteed, without bringing in theorem provers, and spending
dramatically more time than the library took to write.

A best-effort version of this will be to generate a very large
set of additional test data, tantamount to fuzzing the library.
I need to figure out how to add a flag to make running these tests
optional, because testing already takes several seconds, and the
proposed test set will add considerable time to that.