Coverage complete
Only lines which don't trigger are the ones which can't: three
unreachables in runeset, and the expect(false) conditions in
test-runeset.

It would still be good to generate many additional megabytes of
extended test data, to put behind a feature flag.  I'd also like
to fuzz set creation, to see if those unreachables are actually
unreachables, or if I just happen to be unable to come up with
circumstances which would trigger them.

Other than that, and some supporting functionality which I'd like
to add to the library, runeset is complete, and performing according
to the design and requirements.

What a ride.