Add conditional flag for extra tests
I'm tempted to really load the puppy up but at some point it's
actually excessive.

Not because random data guarantees that the code is correct, but
because it's random, and the search space is ridonkulous.  I'd
like to find a way to fuzz the code using something which will
actively hunt for hostile input, but the blogs I've read on fuzzing
Zig have put me off that path for now.