Continue Rune iterator
There's an almost comical amount of code involved here, considering
how dubious it is that anyone will in fact find this part of the
library useful.

Arguably, this is purely art, for art's sake.  I want runeset to
be complete, and after this, and of course, the tests to go with
it, it will be.