Isolate the interesting case
What makes this interesting is that the rune appears to be valid,
but it isn't matching on the set. Unfortunately, that makes the
location of the defect pretty unclear, because it actually ends
up looking like the iterator isn't at fault, even though the
rune is in the test string, which has been tried thousands of times
now against the set.  Just got to keep drilling down...