Reverse order of Rune fields
Finally hit me that you can't lexically sort it without the
Karpinski layout.  The proposal I had in mind for Julia would
mean that the location of the first byte depends on how many
bytes there are, so I was wrong, it's actually not a better choice,
certainly not for Zig, where we can define the bytes as fields.

I mean, it would be *possible* to do it, but we'd lose a nice
property, although we'd get ASCII-equivalence in return.  But the
tradeoff here is not at all worth it.