Solve overlong encoding issues
The Rune contract is that surrogates are conformant, but we don't allow
overlong encodings from API calls.  The first version allowed both.

This makes slice conversion more complex, but simplifies the conformant
version of isScalarValue.  I'm not 100% that we're striking the correct
balance here, but, let's say 90%.  It will do.