Minor efficiency tweak
The `lowerThan` test is optional because it doubles as a test for
membership, so we can avoid checking twice by using it as such a test.

Pure win, since it checks membership first, so the no-match path is the
same speed as it was, and we get a bit of a boost for match path.