Draft of vectorized match function, ztap changes
This has an untested, un-benchmarked version of cursor matching,
which uses a Vector to try and provoke a tight loop to keep a
SIMD register filled with the head of the RuneSet, distributed
so that a byte can be directly bitshifted onto another register
and compared accordingly.  If it works we get single-conditional
head fails.